Refactored to split out tests into different suites
