+++ src/services/order_watcher_service.ts
@@ -20,7 +20,7 @@ export class OrderWatcherService {
         // 1. Get orders from local cache
         const signedOrderModels = (await this._connection.manager.find(SignedOrderEntity)) as Array<
             Required<SignedOrderEntity>
-        >;
+            >;
         const signedOrders = signedOrderModels.map(orderUtils.deserializeOrder);
 
         // 2. Get orders from Mesh
